Add extremely crude 'collection' support
The idea of the collection is that it's a list of notes you build as you
browse around. This list can then be used to generate a link list.

If this feels nice I can build on this and add some way to visualise
which notes are in the list etc.