Fix wikilink hack with more hacks
Since the Glamour parser doesn't understand wikilinks as anything other
than normal text, it became part of the text following a link. When
later being replaced with something that adds and resets the style, the
following text became unstyled.

<textcolor><linkcolor>[[link]]<reset> rest of text
     ^----------------------------------^

The text color for "rest of text" is set before the link color is added,
which then resets the colors.

With this hack

<linkcolor>[[link]]<reset><italic>qwq<reset><textcolor> rest of text.
           This is being removed --^